FIXED ALL WARNINGS YIPPIII 💯
im starving.
but if you still need to go back. Check on all Static Meshes and enable the Nanite, bc i had to gotrough everyone and diable it. Then also look into some material if it is set to Masked -> change it back to Translucent (bc i made it masked) and then there were some rocks set to 0,0,0 , which caused an issue, so i set them all to 1 and changed the mesh to be 1x1x1xsqare mesh, bc if i would have deleted them it would have given an error. Fr if you guys work on stuff dont leave 0,0,0 objects around hahaha , the last hting would be the Niagarar debug ermitter, but i have reset it back to normal.